+++ src/domains/auth/datasources/withdrawApi.ts
@@ -1,8 +1,9 @@
 import apiClient from "@/lib/apis/apiClient";
 import { getServerApiUrl } from "@/lib/utils/api";
 
-const fetchWithdraw = async (accessToken: string): Promise<void> => {
+const withdrawApi = async (accessToken: string): Promise<void> => {
   const endpoint: string = "auth/withdraw";
+
   await apiClient({
     url: getServerApiUrl(endpoint),
     options: {
@@ -14,4 +15,4 @@ const fetchWithdraw = async (accessToken: string): Promise<void> => {
   });
 };
 
-export default fetchWithdraw;
+export default withdrawApi;
