+++ src/domains/auth/usecases/signOut.ts
@@ -6,7 +6,7 @@ import {
 } from "@/lib/cookie/accessToken";
 import { deleteRefreshTokenFromCookie } from "@/lib/cookie/refreshToken";
 
-import fetchSignOut from "@/domains/auth/apis/fetchSignOut";
+import fetchSignOut from "@/domains/auth/repositories/fetchSignOut";
 
 const signOut = async (): Promise<void> => {
   const storedAccessToken: string | null = await getAccessTokenFromCookie();
